Use linked list instead of array for instruction queue
The problem of array -- e.g. std::vector or SmallVector -- is that when
the size grows over its capacity, it will re-allocate memory and undergo
a time-consuming data migration. This will cause significant performance
degradation when the number of instructions grows really fast.
A quick fix for this is simply using linked list (e.g. std::list)